* CC: work around a failure when handling stringize operator.
We should not run stringize operation when the hash char "#" is inside a
string literal. This is just a work around, because we just test whether
a non-space string after # is an alphabetic char. It fixes the failure of
test file cc_macro_expansion_stringize.cpp. Discussions can be found:
http://forums.codeblocks.org/index.php/topic,19791.msg137040.html#msg137040