* CC: fix SF #198: document parsing error with doxygen block comment.
A special member variable m_ReadingMacroDefinition is introduced, so that
when reading the doxygen comments(the ones after Token), we translate
their kind as the one before Token.

int aaa;

Before: the "Mask A..." was wrongly attached to the Token aaa, since
when reading the "Mask A...", the Token FOO is not constructed.

See discussions in: https://sourceforge.net/p/codeblocks/tickets/198/